Testing Infra Cleanup (#1822)
* cleanup tests

* move around fixtures

* refactor

* new autouse file

* remove daily tests

* fix import

* reenable daily

* fix test

* parallel test

* remove xtest

* remove remote option for sftp

* launch nightly

* debug remote

* rerun tests

* rerun tests

* set keys

* try to print dummy env

* give secret access

* test secrets

* try to get nightly running

* add target

* revert _target

* try _target again

* try _target again

* try _target again

* purge _target, add keys

* test on local pr

* log superseecret

* fix remote tests

* run full suite

* try new wandb key

* try doctest

* reneable full suite

* debug daily

* name actions

* change to skip

* add logs

* add logs

* add logs

* add logs

* remove logs

* raise timeout

* remove timeout

* debug log

* remove nightly

* fix tests

* comments
