+++ .github/workflows/pytest.yaml
@@ -25,23 +25,23 @@ jobs:
         include:
           - name: 'cpu-3.8-1.11'
             container: mosaicml/pytorch:1.11.0_cpu-python3.8-ubuntu20.04
-            markers: 'not daily and not gpu and not vision and not remote and not doctest'
+            markers: 'not daily and not remote and not gpu and not vision and not doctest'
             pytest_command: 'coverage run -m pytest'
           - name: 'cpu-3.9-1.12'
             container: mosaicml/pytorch:1.12.1_cpu-python3.9-ubuntu20.04
-            markers: 'not daily and not gpu and not vision and not remote and not doctest'
+            markers: 'not daily and not remote and not gpu and not vision and not doctest'
             pytest_command: 'coverage run -m pytest'
           - name: 'cpu-3.10-1.13'
             container: mosaicml/pytorch:1.13.0_cpu-python3.10-ubuntu20.04
-            markers: 'not daily and not gpu and not vision and not remote and not doctest'
+            markers: 'not daily and not remote and not gpu and not vision and not doctest'
             pytest_command: 'coverage run -m pytest'
           - name: 'cpu-vision'
             container: mosaicml/pytorch_vision:1.12.1_cpu-python3.9-ubuntu20.04
-            markers: 'not daily and not gpu and vision and not remote and not doctest'
+            markers: 'not daily and not remote and not gpu and vision and not doctest'
             pytest_command: 'coverage run -m pytest'
           - name: 'cpu-doctest'
             container: mosaicml/pytorch_vision:1.12.1_cpu-python3.9-ubuntu20.04
-            markers: 'not daily and not gpu and not vision and not remote and doctest'
+            markers: 'not daily and not remote and not gpu and not vision and doctest'
             pytest_command: 'coverage run -m pytest tests/test_docs.py'
     name: ${{ matrix.name }}
     container: ${{ matrix.container }}
