+++ tests/cli/test_cli.py
@@ -10,7 +10,6 @@ import pytest
 import composer
 
 
-@pytest.mark.daily
 @pytest.mark.parametrize('args', [
     ['composer', '--version'],
     [sys.executable, '-m', 'composer', '--version'],
@@ -19,4 +18,4 @@ import composer
 ])
 def test_cli_version(args: List[str]):
     version_str = subprocess.check_output(args, text=True)
-    assert version_str == f'composer {composer.__version__}\n'
+    assert version_str == f'MosaicML Composer {composer.__version__}\n'
