+++ tests/loggers/test_in_memory_logger.py
@@ -6,8 +6,6 @@ import numpy as np
 from composer.core import State, Time, Timestamp
 from composer.loggers import InMemoryLogger, Logger
 
-#from tests.fixtures.new_fixtures import empty_logger, minimal_state
-
 
 def test_in_memory_logger(dummy_state: State):
     in_memory_logger = InMemoryLogger()
